+++ src/index.html
@@ -38,56 +38,11 @@
 
     <body>
         <!-- Service worker to check for an update -->
-        <script>
-            // Check that service workers are supported
-            if ('serviceWorker' in window.navigator) {
-                // Use the window load event to keep the page load performant
-                window.addEventListener('load', () => {
-                    window.navigator.serviceWorker
-                        .register(`${window.location.origin}/sw.js`)
-                        .then((reg) => {
-                            setInterval(() => {
-                                reg.update();
-                            }, 10 * 60 * 1000);
-                        });
-                });
-                window.navigator.serviceWorker.addEventListener(
-                    'controllerchange',
-                    function() {
-                        var result = window.confirm(
-                            'A new version available! Do you want to reload?',
-                        );
-                        if (result) {
-                            window.location.reload();
-                        }
-                    },
-                );
-            }
-        </script>
+        <script src="js/checkingServiceWorkers.js">
+            // Check that service workers are supported</script>
 
         <!-- Google Analytics -->
-        <script>
-            (function(i, s, o, g, r, a, m) {
-                i['GoogleAnalyticsObject'] = r;
-                (i[r] =
-                    i[r] ||
-                    function() {
-                        (i[r].q = i[r].q || []).push(arguments);
-                    }),
-                    (i[r].l = 1 * new Date());
-                (a = s.createElement(o)), (m = s.getElementsByTagName(o)[0]);
-                a.async = 1;
-                a.src = g;
-                m.parentNode.insertBefore(a, m);
-            })(
-                window,
-                document,
-                'script',
-                '//www.google-analytics.com/analytics.js',
-                'ga',
-            );
-            ga('create', 'UA-75111209-1', 'auto');
-            ga('send', 'pageview');
+        <script src="js/googleAnalytics.js">
         </script>
 
         <!-- Navbar -->
