+++ src/js/main.js
@@ -1,6 +1,7 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import '@fortawesome/fontawesome-free/css/all.min.css';
 import '../css/main.scss';
+import '../css/navBar.scss'
 
 import $ from 'jquery';
 import 'bootstrap';
